warningを減らしたのみ

+++ GameChara.cpp
@@ -35,10 +35,10 @@ void GameChara::Draw()
 
 bool GameChara::IsMyRectHit(RectF _rect)
 {
-	float wAB = rect_.w / 2.0 + _rect.w / 2.0;
-	float hAB = rect_.h / 2.0 + _rect.h / 2.0;
-	float distx = abs(rect_.center().x - _rect.center().x);
-	float disty = abs(rect_.center().y - _rect.center().y);
+	float wAB = (float)(rect_.w / 2.0 + _rect.w / 2.0);
+	float hAB = (float)(rect_.h / 2.0 + _rect.h / 2.0);
+	double distx = abs(rect_.center().x - _rect.center().x);
+	double disty = abs(rect_.center().y - _rect.center().y);
 	if (wAB > distx && hAB > disty)
 		return true;
 	else
